1.3.0: cython support for speed boost
